attempts at querying.
Hi! I haven't added to master because this is still experimental. 
I discovered AlertsTipster which is useful for the bit where we are asked to report running time, when not using an inverted index.
That basically does it. 
The queries are one big txt file, it's xml though. We need something that collects all nodes named <title>  - there should be 40 of them, and that will basically be the list of queries!