term-fq model
simple term-fq model. based on a tokenised query, it hashes the query,
then it screens for the union of documents that contain at least one
token from the query using the inverted index, then it goes over each
set in that document set and calculates term-idf for each token in the
query, it then sums these term-idfs together and ranks according to
largest sum, it keeps the top 100 documents for the ranking.